+++ browsertime-master/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Browsertime changelog
+
 UNRELEASED
+-------------------------
+### Added
+* Improved detection of http2 pushed assets for Chrome. #261
+
+version 1.0.0-beta.25 2017-02-15
+-------------------------
+### Added
+* Make it possible to include response bodies in the HAR from Firefox using --firefox.includeResponseBodies [#264](https://github.com/sitespeedio/browsertime/pull/264/)
+* Set Firefox preferences through the CLI [#268](https://github.com/sitespeedio/browsertime/issues/268)
+
+### Fixed
+* Add check if entry is not undefined and request exists when creating the Chrome HAR fixes[#272](https://github.com/sitespeedio/browsertime/issues/272)
+
+version 1.0.0-beta.24 2017-02-10
 -------------------------
 ### Added
 * New option `--userTimingWhitelist` to pass a whitelist regex for filtering userTimings from results
@@ -9,6 +24,7 @@ UNRELEASED
 * Simpliefied check in Chrome if a response is pushed (only use pushStart)
 * Renamed _was_push to _was_pushed as following same name standard as WPT if an entry is HTTP/2 pushed
 * Aligned priority information in HAR with WebPageTest.
+* Upgraded to Geckodriver 0.14.0
 
 version 1.0.0-beta.23 2017-01-10
 -------------------------